Feature/order (#3)
* Adicionado Pacote do Order

- Domain
- Infra
- Services
- API

* Ajustes de Classes para melhor legibilidade e organização de responsabilidades

* Projeto do orders

* Ajustes de serviços e testes manuais;

* Ajustes para execução

* Teste total de integração.

* Ajuste de regra;

* Criado primeiro caso de testes.

+++ services/vehicle/Vehicle.Services/Service/MotoService.cs
@@ -50,6 +50,15 @@ namespace Vehicle.Services.Service
 
         public async Task Remove(long id)
         {
+            MotoFilters filters = new()
+            {
+                Id = id,
+                AllLocated = false,
+                Located = true,
+            };
+            var _hasMoto = await _motoRepository.GetAll(filters);
+            if (_hasMoto == null || _hasMoto.Count() == 0) throw new PersonalizeExceptions("It is not possible to delete a rental vehicle");
+
             await _motoRepository.Delete(id);
         }
